HH-218702 move TransactionalScope to nab-jdbc module
