+++ nab-tests/src/test/java/ru/hh/nab/jpa/transaction/DataSourceContextTest.java
@@ -8,7 +8,6 @@ import org.junit.jupiter.api.Test;
 import org.slf4j.MDC;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.transaction.support.TransactionTemplate;
-import ru.hh.nab.hibernate.transaction.TransactionalScope;
 import static ru.hh.nab.jdbc.common.DataSourceType.MASTER;
 import static ru.hh.nab.jdbc.common.DataSourceType.READONLY;
 import static ru.hh.nab.jdbc.common.DataSourceType.SLOW;
@@ -18,6 +17,7 @@ import static ru.hh.nab.jdbc.routing.DataSourceContext.onSlowReplica;
 import static ru.hh.nab.jdbc.routing.DataSourceContextUnsafe.MDC_KEY;
 import static ru.hh.nab.jdbc.routing.DataSourceContextUnsafe.getDataSourceName;
 import static ru.hh.nab.jdbc.routing.DataSourceContextUnsafe.setDefaultMDC;
+import ru.hh.nab.jdbc.transaction.TransactionalScope;
 import ru.hh.nab.jpa.JpaTestConfig;
 import ru.hh.nab.testbase.jpa.JpaTestBase;
